move getMissingChanges and getClock to clocks.ts
